+++ addons/docs/src/lib/docgen/createPropDef.ts
@@ -1,8 +1,10 @@
 import { isNil } from 'lodash';
-import { PropDef } from '@storybook/components';
-import { TypeSystem, DocgenInfo, DocgenType } from './types';
+import { PropDef, PropDefaultValue } from '@storybook/components';
+import { TypeSystem, DocgenInfo, DocgenType, DocgenPropDefaultValue } from './types';
 import { JsDocParsingResult } from '../jsdocParser';
-import { createDefaultValue } from './createDefaultValue';
+import { createSummaryValue } from '../utils';
+import { createFlowPropDef } from './flow/createPropDef';
+import { isDefaultValueBlacklisted } from './utils/defaultValue';
 
 export type PropDefFactory = (
   propName: string,
@@ -10,30 +12,36 @@ export type PropDefFactory = (
   jsDocParsingResult?: JsDocParsingResult
 ) => PropDef;
 
+function createDefaultValue(defaultValue: DocgenPropDefaultValue): PropDefaultValue {
+  if (!isNil(defaultValue)) {
+    const { value } = defaultValue;
+
+    if (!isDefaultValueBlacklisted(value)) {
+      return createSummaryValue(value);
+    }
+  }
+
+  return null;
+}
+
 function createBasicPropDef(name: string, type: DocgenType, docgenInfo: DocgenInfo): PropDef {
   const { description, required, defaultValue } = docgenInfo;
 
   return {
     name,
-    type: { summary: type.name },
+    type: createSummaryValue(type.name),
     required,
     description,
     defaultValue: createDefaultValue(defaultValue),
   };
 }
 
-function createPropDef(
-  name: string,
-  type: DocgenType,
-  docgenInfo: DocgenInfo,
-  jsDocParsingResult: JsDocParsingResult
-): PropDef {
-  const propDef = createBasicPropDef(name, type, docgenInfo);
-
+function applyJsDocResult(propDef: PropDef, jsDocParsingResult: JsDocParsingResult): PropDef {
   if (jsDocParsingResult.includesJsDoc) {
     const { description, extractedTags } = jsDocParsingResult;
 
     if (!isNil(description)) {
+      // eslint-disable-next-line no-param-reassign
       propDef.description = jsDocParsingResult.description;
     }
 
@@ -41,6 +49,7 @@ function createPropDef(
     const hasReturns = !isNil(extractedTags.returns) && !isNil(extractedTags.returns.type);
 
     if (hasParams || hasReturns) {
+      // eslint-disable-next-line no-param-reassign
       propDef.jsDocTags = {
         params:
           hasParams &&
@@ -53,36 +62,28 @@ function createPropDef(
   return propDef;
 }
 
-export const javaScriptFactory: PropDefFactory = (
-  propName: string,
-  docgenInfo: DocgenInfo,
-  jsDocParsingResult?: JsDocParsingResult
-) => {
-  return createPropDef(propName, docgenInfo.type, docgenInfo, jsDocParsingResult);
+export const javaScriptFactory: PropDefFactory = (propName, docgenInfo, jsDocParsingResult) => {
+  const propDef = createBasicPropDef(propName, docgenInfo.type, docgenInfo);
+
+  return applyJsDocResult(propDef, jsDocParsingResult);
 };
 
-export const tsFactory: PropDefFactory = (
-  propName: string,
-  docgenInfo: DocgenInfo,
-  jsDocParsingResult?: JsDocParsingResult
-) => {
-  return createPropDef(propName, docgenInfo.tsType, docgenInfo, jsDocParsingResult);
+export const tsFactory: PropDefFactory = (propName, docgenInfo, jsDocParsingResult) => {
+  const propDef = createBasicPropDef(propName, docgenInfo.tsType, docgenInfo);
+
+  return applyJsDocResult(propDef, jsDocParsingResult);
 };
 
-export const flowFactory: PropDefFactory = (
-  propName: string,
-  docgenInfo: DocgenInfo,
-  jsDocParsingResult?: JsDocParsingResult
-) => {
-  return createPropDef(propName, docgenInfo.flowType, docgenInfo, jsDocParsingResult);
+export const flowFactory: PropDefFactory = (propName, docgenInfo, jsDocParsingResult) => {
+  const propDef = createFlowPropDef(propName, docgenInfo);
+
+  return applyJsDocResult(propDef, jsDocParsingResult);
 };
 
-export const unknownFactory: PropDefFactory = (
-  propName: string,
-  docgenInfo: DocgenInfo,
-  jsDocParsingResult?: JsDocParsingResult
-) => {
-  return createPropDef(propName, { name: 'unknown' }, docgenInfo, jsDocParsingResult);
+export const unknownFactory: PropDefFactory = (propName, docgenInfo, jsDocParsingResult) => {
+  const propDef = createBasicPropDef(propName, { name: 'unknown' }, docgenInfo);
+
+  return applyJsDocResult(propDef, jsDocParsingResult);
 };
 
 export const getPropDefFactory = (typeSystem: TypeSystem): PropDefFactory => {