+++ examples/official-storybook/stories/addon-info/react-docgen.stories.js
@@ -13,7 +13,7 @@ export default {
   decorators: [withInfo],
 };
 
-export const commentsFromPropTypeDeclarations = () => (
+export const CommentsFromPropTypeDeclarations = () => (
   <DocgenButton
     onClick={action('clicked')}
     label="Docgen Button"
@@ -33,7 +33,7 @@ export const commentsFromPropTypeDeclarations = () => (
   />
 );
 
-commentsFromPropTypeDeclarations.story = {
+CommentsFromPropTypeDeclarations.story = {
   name: 'Comments from PropType declarations',
 
   parameters: {
@@ -42,11 +42,11 @@ commentsFromPropTypeDeclarations.story = {
   },
 };
 
-export const commentsFromFlowDeclarations = () => (
+export const CommentsFromFlowDeclarations = () => (
   <FlowTypeButton onClick={action('clicked')} label="Flow Typed Button" />
 );
 
-commentsFromFlowDeclarations.story = {
+CommentsFromFlowDeclarations.story = {
   name: 'Comments from Flow declarations',
 
   parameters: {
@@ -55,11 +55,11 @@ commentsFromFlowDeclarations.story = {
   },
 };
 
-export const commentsFromComponentDeclaration = () => (
+export const CommentsFromComponentDeclaration = () => (
   <BaseButton onClick={action('clicked')} label="Button" />
 );
 
-commentsFromComponentDeclaration.story = {
+CommentsFromComponentDeclaration.story = {
   name: 'Comments from component declaration',
 
   parameters: {
@@ -68,11 +68,11 @@ commentsFromComponentDeclaration.story = {
   },
 };
 
-export const commentsFromNamedExportComponentDeclaration = () => (
+export const CommentsFromNamedExportComponentDeclaration = () => (
   <NamedExportButton onClick={action('clicked')} label="Button" />
 );
 
-commentsFromNamedExportComponentDeclaration.story = {
+CommentsFromNamedExportComponentDeclaration.story = {
   name: 'Comments from named export component declaration',
 
   parameters: {