+++ examples/polymer-cli/src/stories/addon-actions.stories.js
@@ -7,22 +7,22 @@ export default {
   title: 'Addon/Actions',
 };
 
-export const actionOnly = () => {
+export const ActionOnly = () => {
   const el = document.createElement('simple-button');
   el.addEventListener('click', action('log1'));
   return el;
 };
 
-actionOnly.story = {
+ActionOnly.story = {
   name: 'Action only',
 };
 
-export const actionAndMethod = () => {
+export const ActionAndMethod = () => {
   const el = document.createElement('simple-button');
   el.addEventListener('click', e => action('log2')(e.target));
   return el;
 };
 
-actionAndMethod.story = {
+ActionAndMethod.story = {
   name: 'Action and method',
 };