+++ lib/components/src/blocks/PropsTable/PropJsDoc.tsx
@@ -2,29 +2,68 @@ import React, { FC } from 'react';
 import { styled } from '@storybook/theming';
 import { isNil } from 'lodash';
 import { JsDocTags } from './PropDef';
+import { codeCommon } from '../../typography/shared';
 
 interface PropJsDocProps {
   tags: JsDocTags;
 }
 
-const TBody = styled.tbody({ boxShadow: 'none !important' });
+export const Table = styled.table(({ theme }) => ({
+  '&&': {
+    // Escape default table styles
+    borderCollapse: 'collapse',
+    borderSpacing: 0,
+    border: 'none',
 
-const Cell = { paddingTop: '0 !important', paddingBottom: '0 !important' };
+    tr: {
+      border: 'none !important',
+      background: 'none',
+    },
 
-const Name = styled.td({
-  ...Cell,
-});
+    'td, th': {
+      padding: 0,
+      border: 'none',
+      width: 'auto!important',
+    },
+    // End escape
 
-const Desc = styled.td({
-  ...Cell,
-  width: 'auto !important',
-});
+    marginTop: '0',
+    marginBottom: '0',
+
+    'th:first-of-type, td:first-of-type': {
+      paddingLeft: 0,
+    },
+
+    'th:last-of-type, td:last-of-type': {
+      paddingRight: 0,
+    },
+
+    td: {
+      paddingTop: '0',
+      paddingBottom: '4px',
+
+      '&:not(:first-of-type)': {
+        paddingLeft: 10,
+        paddingRight: 0,
+      },
+    },
+
+    tbody: {
+      boxShadow: 'none',
+      border: 'none',
+    },
+
+    code: codeCommon({ theme }),
+
+    '& code': {
+      lineHeight: '18px',
+      margin: 0,
+      display: 'inline-block',
+    },
+  },
+}));
 
 export const PropJsDoc: FC<PropJsDocProps> = ({ tags }) => {
-  if (isNil(tags)) {
-    return null;
-  }
-
   const params = (tags.params || []).filter(x => x.description);
   const hasDisplayableParams = params.length !== 0;
   const hasDisplayableReturns = !isNil(tags.returns) && !isNil(tags.returns.description);
@@ -34,28 +73,28 @@ export const PropJsDoc: FC<PropJsDocProps> = ({ tags }) => {
   }
 
   return (
-    <table>
-      <TBody>
+    <Table>
+      <tbody>
         {hasDisplayableParams &&
           params.map(x => {
             return (
               <tr key={x.name}>
-                <Name>
+                <td>
                   <code>{x.name}</code>
-                </Name>
-                <Desc>{x.description}</Desc>
+                </td>
+                <td>{x.description}</td>
               </tr>
             );
           })}
         {hasDisplayableReturns && (
           <tr key="returns">
-            <Name>
+            <td>
               <code>Returns</code>
-            </Name>
-            <Desc>{tags.returns.description}</Desc>
+            </td>
+            <td>{tags.returns.description}</td>
           </tr>
         )}
-      </TBody>
-    </table>
+      </tbody>
+    </Table>
   );
 };